Fix tests by splitting integration and unit tests.
Use JDK8.
