+++ .travis.yml
@@ -1,2 +1,6 @@
 language: java
+
+jdk:
+  - oraclejdk8
+
 script: mvn verify
